requirements.txt
telethon==1.26.0
python-dotenv==0.21.0
fastapi==0.88.0
uvicorn==0.20.0
